🚚 [projects] Move function `generate` from `services` to `projects`
