+++ src/cutty/projects/generate.py
