+++ src/cutty/services/create.py
@@ -4,7 +4,7 @@ from collections.abc import Sequence
 from typing import Optional
 
 from cutty.projects.create import creategitrepository
-from cutty.services.generate import generate
+from cutty.projects.generate import generate
 from cutty.services.loadtemplate import loadtemplate
 from cutty.templates.domain.bindings import Binding
